requirements.txt
Django==3.0.3
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.9
pytz==2019.3
requests==2.23.0
urllib3==1.25.9
